Chore: Split Gatsby Node.js code
