+++ gatsby-node.js
@@ -1,103 +1,7 @@
-const { resolve: resolvePath } = require('path');
+const { createPages, onCreateNode, onCreatePage } = require('./gatsby');
 
-const { createRemoteFileNode } = require('gatsby-source-filesystem');
+exports.createPages = createPages;
 
-const caseDetailTemplate = resolvePath('./src/cases/detail-template.jsx');
-const jobDetailTemplate = resolvePath('./src/jobs/detail-template.jsx');
-const postDetailTemplate = resolvePath('./src/posts/detail-template.jsx');
+exports.onCreateNode = onCreateNode;
 
-const WORDPRESS_IMAGES = [
-  { type: 'wordpress__wp_members', fields: ['portret'] },
-];
-
-exports.createPages = async function createPages({
-  actions: { createPage },
-  graphql,
-}) {
-  function createWPPages(nodes, component) {
-    nodes.forEach(({ path, slug }) => {
-      createPage({
-        path,
-        component,
-        context: { slug },
-      });
-    });
-  }
-  const {
-    data: { allWordpressWpCases, allWordpressWpJobs, allWordpressPost },
-    errors,
-  } = await graphql(
-    `
-      query {
-        allWordpressWpCases(filter: { status: { eq: "publish" } }) {
-          nodes {
-            path
-            slug
-          }
-        }
-        allWordpressWpJobs(filter: { status: { eq: "publish" } }) {
-          nodes {
-            path
-            slug
-          }
-        }
-        allWordpressPost(filter: { status: { eq: "publish" } }) {
-          nodes {
-            path
-            slug
-          }
-        }
-      }
-    `,
-  );
-  if (errors) {
-    throw errors;
-  }
-  createWPPages(allWordpressPost.nodes, postDetailTemplate);
-  createWPPages(allWordpressWpCases.nodes, caseDetailTemplate);
-  createWPPages(allWordpressWpJobs.nodes, jobDetailTemplate);
-};
-
-exports.onCreateNode = async function onCreateNode({
-  actions: { createNode, createNodeField },
-  cache,
-  createNodeId,
-  node,
-  store,
-}) {
-  /* eslint no-param-reassign: "off" */
-  await Promise.all(
-    WORDPRESS_IMAGES.map(async ({ type, fields }) => {
-      if (node.internal.type === type) {
-        await Promise.all(
-          fields.map(async (field) => {
-            const fileNode = await createRemoteFileNode({
-              cache,
-              createNode,
-              createNodeId,
-              parentNodeId: node.id,
-              store,
-              url: node[field].url,
-            });
-            createNodeField({ node, name: field, value: fileNode });
-            node[`${field}___NODE`] = fileNode.id;
-          }),
-        );
-      }
-    }),
-  );
-};
-
-exports.onCreatePage = ({ page, actions: { createPage, deletePage } }) => {
-  if (!page.context.intl || page.context.language) {
-    return;
-  }
-  deletePage(page);
-  createPage({
-    ...page,
-    context: {
-      ...page.context,
-      language: page.context.intl.language,
-    },
-  });
-};
+exports.onCreatePage = onCreatePage;
