+++ gatsby/index.js
@@ -1,0 +1,5 @@
+const createPages = require('./create-pages');
+const onCreateNode = require('./on-create-node');
+const onCreatePage = require('./on-create-page');
+
+module.exports = { createPages, onCreateNode, onCreatePage };
